fix(gateway): restore suspension CI gates
