+++ src/auto-reply/reply/queue.drain-restart.test.ts
@@ -257,7 +257,6 @@ describe("followup queue drain restart after idle window", () => {
       import.meta.url,
       "./queue/enqueue.js?scope=restart-b",
     );
-    const { clearSessionQueues } = await import("./queue.js");
     const key = `test-idle-window-cross-module-${Date.now()}`;
     const calls: FollowupRun[] = [];
     const settings: QueueSettings = { mode: "followup", debounceMs: 0, cap: 50 };
@@ -522,7 +521,6 @@ describe("followup queue drain restart after idle window", () => {
       setImmediate(resolve);
     });
 
-    const { clearSessionQueues } = await import("./queue.js");
     clearSessionQueues([key]);
 
     enqueueFollowupRun(key, createRun({ prompt: "after-clear" }), settings);
