+++ src/gateway/local-request-context.ts
@@ -11,7 +11,7 @@ import {
 import { NodeRegistry } from "./node-registry.js";
 import type { ChannelRuntimeSnapshot } from "./server-channel-runtime.types.js";
 import { createChatRunEntry, type ChatRunEntry } from "./server-chat-state.js";
-import type { GatewayCronServiceContract } from "./server-cron.js";
+import type { GatewayCronServiceContract } from "./server-cron.types.js";
 import type { GatewayRequestContext } from "./server-methods/types.js";
 
 // Embedded/local agent calls need enough GatewayRequestContext to reuse server
