+++ src/gateway/server-cron-lazy.test.ts
@@ -4,7 +4,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import type { CliDeps } from "../cli/deps.types.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import type { GatewayCronServiceContract, GatewayCronState } from "./server-cron.js";
+import type { GatewayCronServiceContract, GatewayCronState } from "./server-cron.types.js";
 
 const hoisted = vi.hoisted(() => {
   let state: unknown;
