+++ src/gateway/server-reload-handlers.ts
@@ -40,7 +40,8 @@ import type { ChannelHealthMonitor } from "./channel-health-monitor.js";
 import type { ChannelKind } from "./config-reload-plan.js";
 import { startGatewayConfigReloader, type GatewayReloadPlan } from "./config-reload.js";
 import { resolveHooksConfig } from "./hooks.js";
-import { buildGatewayCronService, type GatewayCronState } from "./server-cron.js";
+import { buildGatewayCronService } from "./server-cron.js";
+import type { GatewayCronState } from "./server-cron.types.js";
 import { applyGatewayLaneConcurrency } from "./server-lanes.js";
 import { markGatewayModelCatalogStaleForReload } from "./server-model-catalog.js";
 import type { GatewayConfigReloaderHandle } from "./server-runtime-handles.js";
